requirements.txt
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
django-cors-headers==2.4.0
django-taggit==0.23.0
django-taggit-serializer==0.1.7
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.3.0
faker==0.9.1
gunicorn==19.9.0
httpie==0.9.9
idna==2.7
Pillow==5.2.0
psycopg2-binary==2.7.5
Pygments==2.2.0
PyJWT==1.6.4
pytest==3.8.1
python-decouple==3.1
pytest==3.8.1
pytz==2018.5
requests==2.19.1
six==1.11.0
stripe==2.8.1
urllib3==1.23
whitenoise==4.1
